REFACTOR: Use errors.py exceptions
