+++ src/ansys/aedt/core/generic/configurations.py
@@ -32,7 +32,7 @@ import ansys.aedt.core
 from ansys.aedt.core import __version__
 from ansys.aedt.core.application.variables import decompose_variable_value
 from ansys.aedt.core.generic.data_handlers import _arg2dict
-from ansys.aedt.core.generic.general_methods import GrpcApiError
+from ansys.aedt.core.generic.errors import GrpcApiError
 from ansys.aedt.core.generic.general_methods import generate_unique_folder_name
 from ansys.aedt.core.generic.general_methods import generate_unique_name
 from ansys.aedt.core.generic.general_methods import open_file
