+++ src/ansys/aedt/core/generic/grpc_plugin_dll_class.py
@@ -34,7 +34,7 @@ import socket
 import time
 import types
 
-from ansys.aedt.core.generic.general_methods import GrpcApiError
+from ansys.aedt.core.generic.errors import GrpcApiError
 from ansys.aedt.core.generic.general_methods import _retry_ntimes
 from ansys.aedt.core.generic.general_methods import inclusion_list
 from ansys.aedt.core.generic.general_methods import settings
